constraint_dc_current_mag_gwye_gwye_xf
Made the relaxation of constraint_dc_current_mag_gwye_gwye_xf no longer the default implementation.  Now, if a user tries to use a model for which there is no implementation constraint_dc_current_mag_gwye_gwye_xf, the code will produce an error message to indicate that such an implementation needs to be created to force the user to make a decision about what version of the constraint is appropriate for the model.